+++ src/graphql/resolvers/user/index.ts
@@ -1,0 +1,3 @@
+export { default as registration } from './registration';
+export { default as login } from './login';
+export { default as getOne } from './getOne';