+++ src/prisma/migrations/20210209205541_add_user_last_login/migration.sql
@@ -1,0 +1,17 @@
+-- RedefineTables
+PRAGMA foreign_keys=OFF;
+CREATE TABLE "new_User" (
+    "id" INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
+    "email" TEXT NOT NULL,
+    "password" TEXT NOT NULL,
+    "name" TEXT,
+    "role" INTEGER DEFAULT 0,
+    "lastLogin" DATETIME,
+    "created" DATETIME DEFAULT CURRENT_TIMESTAMP,
+    "updated" DATETIME DEFAULT CURRENT_TIMESTAMP
+);
+INSERT INTO "new_User" ("id", "email", "password", "name", "role", "lastLogin", "created", "updated") SELECT "id", "email", "password", "name", "role", "lastLogin", "created", "updated" FROM "User";
+DROP TABLE "User";
+ALTER TABLE "new_User" RENAME TO "User";
+PRAGMA foreign_key_check;
+PRAGMA foreign_keys=ON;